Move readiness flags to shared module
Copilot-Session: ac86dd2d-0280-4d7b-87a8-5654ad1503a6
