+++ litebox_broker_protocol/src/lib.rs
@@ -17,6 +17,7 @@ pub mod error;
 pub mod event;
 pub mod message;
 pub mod pipe;
+pub mod readiness;
 pub mod wire;
 
 /// Opaque broker object reference handle.
